requirements.txt
Flask
diffusers
torch
Pillow
accelerate 
transformers 
safetensors
gunicorn